Fix Transform calc_matrices method

+++ src/geo/mod.rs
@@ -9,7 +9,7 @@ mod transform;
 * Uses the WGS-84 approximation. The radius at the equator is ~6378137 and at the poles is ~6356752. https://en.wikipedia.org/wiki/World_Geodetic_System#WGS84
 * 6371008.8 is one published "average radius" see https://en.wikipedia.org/wiki/Earth_radius#Mean_radius, or ftp://athena.fsv.cvut.cz/ZFG/grs80-Moritz.pdf p.4
 */
-const EARTH_RADIUS: f32 = 6371008.8;
+const EARTH_RADIUS: f64 = 6371008.8;
 
 pub(crate) use lng_lat::LngLat;
 pub(crate) use lng_lat_bounds::LngLatBounds;
